fix: the ts wrong of card inteface (#35753)
* fix: the ts wrong of card inteface

* fix: add export cardproprs

* feat: add card type test
